chore(openapi2kong): rename package to openapi2kong
improved consistency
