+++ .gitignore
@@ -18,7 +18,7 @@
 dist/
 
 # generated test files
-convertoas3/oas3_testfiles/*.generated.json
+openapi2kong/oas3_testfiles/*.generated.json
 merge/merge_testfiles/*_generated.json
 
 # the binary
