--- convertoas3/convertoas3_test.go
@@ -1,12 +1,0 @@
-package convertoas3_test
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-)
-
-var _ = Describe("convertoas3", func() {
-	Describe("all to do", func() {
-		PIt("still to do", func() {
-		})
-	})
-})
