+++ openapi2kong/oas3_testfiles/01-names-inferred.yaml
