+++ openapi2kong/oas3_testfiles/02-names-set.yaml
