+++ openapi2kong/oas3_testfiles/03a-server-no-trailing-slash.expected.json
