+++ openapi2kong/oas3_testfiles/04-servers-upstream.yaml
