+++ openapi2kong/oas3_testfiles/05-service-defaults-document.yaml
