+++ openapi2kong/oas3_testfiles/06-upstream-defaults-document.yaml
