+++ openapi2kong/oas3_testfiles/06a-upstream-defaults.yaml
