+++ openapi2kong/oas3_testfiles/07-service-defaults-overrides.yaml
