+++ openapi2kong/oas3_testfiles/07a-service-defaults.expected.json
