+++ openapi2kong/oas3_testfiles/08-route-defaults-overrides.expected.json
