+++ openapi2kong/oas3_testfiles/09a-plugins-with-consumers.yaml
