+++ openapi2kong/oas3_testfiles/10-generic-plugins-multi-service.expected.json
