+++ openapi2kong/oas3_testfiles/11-references.yaml
