+++ openapi2kong/oas3_testfiles/12-path-parameter-regex.expected.json
