+++ openapi2kong/oas3_testfiles/15-circular-requestBody-schema.yaml
