+++ openapi2kong/service_test.go
@@ -1,4 +1,4 @@
-package convertoas3
+package openapi2kong
 
 import (
 	"net/url"
